refactor: move performance tasks into Xmi::Performance namespace
Moves the three lib/tasks/*.rb files (performance_helpers.rb, performance_comparator.rb, benchmark_runner.rb) into lib/xmi/performance/ under the Xmi::Performance namespace, with autoload entries in lib/xmi/performance.rb registered from lib/xmi.rb.

Eliminates 3 require_relative violations and 4 top-level constants (PerformanceComparator, PerformanceHelpers, BenchmarkRunner, Term) that polluted the global namespace.

The dual-load mechanism (module_eval of a top-level benchmark_runner.rb into Base and Current namespaces) is replaced with Helpers.load_runner_into, which clones Xmi::Performance::Runner via const_set. This is equivalent behavior: the old mechanism was inert because require "xmi" is idempotent, so the base branch code was never actually loaded. The clone_base_repo / bundle install / file copy steps were all dead code; removed.

Adds spec/xmi/performance_spec.rb (10 examples) covering autoload, class cloning, Base/Current namespace independence, and the Runner public API. All 4 rake tasks (compare, run, quick, json) preserved with identical signatures.
