+++ lib/tasks/performance.rake
@@ -1,34 +1,32 @@
 # frozen_string_literal: true
 
-require_relative "performance_comparator"
-require_relative "benchmark_runner"
+require "xmi"
 
 desc "Run performance benchmarks"
 namespace :performance do
   desc "Compare performance of current branch against base branch (default: main)"
   task :compare do
-    PerformanceComparator.new.run
+    Xmi::Performance::Comparator.new.run
   end
 
   desc "Run benchmarks on current branch only (for development)"
   task :run do
-    runner = BenchmarkRunner.new(run_time: 5)
+    runner = Xmi::Performance::Runner.new(run_time: 5)
     runner.run_benchmarks
   end
 
   desc "Quick benchmark run (faster, less accurate)"
   task :quick do
-    runner = BenchmarkRunner.new(run_time: 2, warmup: 1)
+    runner = Xmi::Performance::Runner.new(run_time: 2, warmup: 1)
     runner.run_benchmarks
   end
 
   desc "Run benchmarks and output as JSON"
   task :json do
     require "json"
-    runner = BenchmarkRunner.new(run_time: 5)
+    runner = Xmi::Performance::Runner.new(run_time: 5)
 
-    # Suppress pretty output, just get results
-    results = runner.send(:run_benchmarks)
+    results = runner.run_benchmarks
 
     output = results.each_with_object({}) do |(label, metrics), h|
       ips = (metrics[:lower] + metrics[:upper]) / 2.0
